add yared task

+++ src/main/java/edu/miu/e_mart/repository/ICustomerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 import edu.miu.e_mart.domain.Customer;
 @Repository
-public interface ICustomerRepository extends CrudRepository<Customer, Long>{
+public interface ICustomerRepository extends CrudRepository<Customer, Integer >{
 
 }
